fix(crash): zlib buffer overflow SIGSEGV + enfeebling nil + pet_hp overflow
Three root causes from today crashes:

1. NetworkBuffer 2500->4096: zlib_compress max_sz=(out_sz-1)*8=19992 bits;
   packet batches with long Huffman codes overflow output buffer, silent
   memory corruption -> SIGSEGV. Pass PScratchBuffer.size() to compress and
   decompress so the buffer and size stay in sync. REQUIRES REBUILD.

2. enfeebling_song.lua:151 nil guard: useEnfeeblingSong indexed pTable[spellId]
   without checking if the spell ID exists; unknown IDs nil-crashed the server.

3. char_stats.pet_hp smallint->int: BST jug pets get 280k+ HP from the overhaul;
   smallint(4) unsigned caps at 65535, causing repeated DB out-of-range errors
   on every char save. Migration in modules/custom/sql/widen_pet_hp_column.sql.

+++ sql/char_stats.sql
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS `char_stats` (
   `pet_id` smallint(3) unsigned NOT NULL DEFAULT '0',
   `pet_type` smallint(3) unsigned NOT NULL DEFAULT '0',
   `pet_level` smallint(3) unsigned NOT NULL DEFAULT '0',
-  `pet_hp` smallint(4) unsigned NOT NULL DEFAULT '0',
+  `pet_hp` int(10) unsigned NOT NULL DEFAULT '0',
   `pet_mp` smallint(4) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`charid`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
